Split SPE fit into new multiple compilation units
