+++ src/calib/CMakeLists.txt
@@ -13,7 +13,8 @@
 # WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
 # A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 
-add_library(calin_calib SHARED spe_fit.cpp pmt_model_pg.cpp)
+add_library(calin_calib SHARED spe_fit.cpp spe_fit_poisson_gaussian.cpp
+spe_fit_general_poisson.cpp pmt_model_pg.cpp)
 target_link_libraries(calin_calib calin_math)
 target_link_libraries(calin_calib ${FFTW_LIBRARIES} ${EXTRA_LIBS})
